pyproject.toml
[project]
name = "livraria"
version = "0.1.0"
description = "Projeto Livraria 2023 - marrcandre"
authors = [
    {name = "renascm_", email = "renatochagas.m@gmail.com"},
]
dependencies = ["django>=4.2.20", "django-restframework>=0.0.1", "djangorestframework>=3.15.2", "django-cors-headers>=4.7.0", "djangorestframework-simplejwt>=5.5.0"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
startapp = "python manage.py startapp"